requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
requests
aiofiles
jinja2
markdown
